Rename util.Predicates package to util.predicate
